LayoutTests/html5lib/generated/run-tests18-write-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
../resources/tests18.dat: PASS
